auth ban/unban. wh bug fixes

+++ controllers/EmployeeController.php
@@ -8,6 +8,8 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\filters\AccessControl;
+use Yii;
+use app\models\Lending;
 
 /**
  * EmployeeController implements the CRUD actions for Employee model.
@@ -126,10 +128,31 @@ class EmployeeController extends Controller
      */
     public function actionDelete($id_employee)
     {
-        $this->findModel($id_employee)->delete();
-
+        // Check if there are any lending records for the employee with type = 1
+        $lendingExist = $this->findLending($id_employee);
+        
+        if (!$lendingExist) {
+            // If no lending records exist, delete the employee
+            $this->findModel($id_employee)->delete();
+            
+            // Set a success notification
+            Yii::$app->session->setFlash('success', 'Employee data deleted successfully.');
+        } else {
+            // Set an error notification if the employee has lending records
+            Yii::$app->session->setFlash('error', 'Employee cannot be deleted. Unit of item still being lent to this employee');
+        }
+        
+        // Redirect to the index page
         return $this->redirect(['index']);
     }
+    
+
+    protected function findLending($id_employee)
+    {
+        // Check if there are any Lending records with the given employee ID and type = 1
+        return Lending::find()->where(['id_employee' => $id_employee, 'type' => 1])->exists();
+    }
+    
 
     /**
      * Finds the Employee model based on its primary key value.
